[TabTray] Remove TabsSessionModel from TabTrayFragment
